requirements.txt
asgiref==3.7.2
certifi==2023.11.17
chardet==3.0.4
Django==4.2.6
django-autoslug==1.9.9
django-taggit==5.0.1
djangorestframework==3.14.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==10.0.1
psycopg2==2.9.9
pytz==2023.3.post1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
